File Formats Bonanza/pyproject.toml
[tool.poetry]
name = "file formats bonanza"
version = "0.1.0"
description = ""
authors = ["KennetTech <64315533+KennetTech@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
